refactor: Controller 객체를 페키지 루트로 이동시켰다.
Controller 가 콘솔 view 들과 도메인에 대한 디펜던시를 갖고 있는 MVC 패턴 적용 중이기에 루트 패키지에 있는 것이 더 적합하다고 판단된다.

참조: https://github.com/next-step/kotlin-blackjack/pull/762#discussion_r1866842515
